FAC-747:feature:Move LoadAnswerRangesPorts to answerrange package
